Add multiple updates and fixes to improve functionality and performance
- Add quality of life updates, including memory and node counting enhancements.
- Restructured the order of Burrows-Wheeler Transform (BWT) and Suffix Array (SA) construction for efficiency.
- Fixed scope issues in createIndex when BIGBWT is not usable.
- Improved management of the writer buffer.
- Enhanced logging functionality in MovePhiReprBP.
- Corrected issues with position calculations and loop conditions in BMove methods.
- Refactored CMakeLists.txt to suppress developer warnings and updated external dependencies.
- Addressed segmentation faults and ensured proper execution return values in the main function.
- Improved installation instructions in the README for better clarity.
- Add installation scripts.
- Integrated divsufsort for handling large texts and 64-bit numbers.

+++ src/fmindex/fmindex.h
@@ -73,7 +73,7 @@ class FMIndex : public IndexInterface {
         inTextMatrix128; // pointer to the current inTextMatrix for many
                          // errors/start positions
 
-    thread_local static std::vector<uint16_t> zerosBuffer;
+    thread_local static std::vector<uint32_t> zerosBuffer;
 
     bool noCIGAR = false; // do not calculate the CIGAR string
 
@@ -282,7 +282,7 @@ class FMIndex : public IndexInterface {
     initializeAndSelectInTextMatrix(length_t nZeros, length_t maxED,
                                     const Substring& pattern) const {
 
-             bool matrix64 = use64Matrix(nZeros, maxED);
+        bool matrix64 = use64Matrix(nZeros, maxED);
 
         IBitParallelED* matrix;
         if (matrix64)
